refactor: split generation studio composables
