+++ app/frontend/src/composables/useGenerationStudio.ts
@@ -1,34 +1,33 @@
-import { computed, onMounted, onUnmounted, ref, watch, type ComputedRef } from 'vue'
+import { onMounted, ref } from 'vue'
 import { storeToRefs } from 'pinia'
 
-import { useActiveJobsApi, useRecentResultsApi, useSystemStatusApi } from '@/composables/apiClients'
+import { useGenerationPersistence } from '@/composables/useGenerationPersistence'
+import { useGenerationUpdates } from '@/composables/useGenerationUpdates'
 import {
   cancelGenerationJob,
   deleteGenerationResult,
-  resolveBackendUrl,
-  resolveGenerationBaseUrl,
   startGeneration as startGenerationRequest,
   toGenerationRequestPayload,
 } from '@/services/generationService'
 import { useAppStore } from '@/stores/app'
 import { useSettingsStore } from '@/stores/settings'
 import type {
-  GenerationCompleteMessage,
-  GenerationErrorMessage,
   GenerationFormState,
   GenerationJob,
-  GenerationProgressMessage,
   GenerationResult,
   NotificationType,
-  ProgressUpdate,
-  SystemStatusPayload,
   SystemStatusState,
-  WebSocketMessage,
 } from '@/types'
+
+const normalizeProgress = (value?: number | null): number => {
+  if (typeof value !== 'number' || Number.isNaN(value)) {
+    return 0
+  }
+  return value <= 1 ? Math.round(value * 100) : Math.round(value)
+}
 
 export const useGenerationStudio = () => {
   const appStore = useAppStore()
-  const { activeJobs, recentResults } = storeToRefs(appStore)
   const settingsStore = useSettingsStore()
   const { backendUrl: configuredBackendUrl } = storeToRefs(settingsStore)
 
@@ -52,253 +51,37 @@ export const useGenerationStudio = () => {
   const showModal = ref(false)
   const selectedResult = ref<GenerationResult | null>(null)
 
-  const websocket = ref<WebSocket | null>(null)
-  const pollInterval = ref<number | null>(null)
-  const isConnected = computed<boolean>(() => websocket.value?.readyState === WebSocket.OPEN)
-
-  const appendWebSocketPath = (path: string): string => {
-    const trimmed = path.replace(/\/+$/, '')
-    if (trimmed) {
-      return `${trimmed}/ws/progress`
-    }
-    return '/api/v1/ws/progress'
-  }
-
-  const resolveWebSocketUrl = (backendUrl?: string | null): string => {
-    const base = resolveGenerationBaseUrl(backendUrl)
-
-    if (/^https?:\/\//i.test(base)) {
-      try {
-        const url = new URL(base)
-        const protocol = url.protocol === 'https:' ? 'wss:' : 'ws:'
-        return `${protocol}//${url.host}${appendWebSocketPath(url.pathname)}`
-      } catch (error) {
-        console.error('Failed to parse backend URL for WebSocket:', error)
-      }
-    }
-
-    const wsPath = appendWebSocketPath(base)
-
-    if (typeof window === 'undefined') {
-      return wsPath
-    }
-
-    const normalizedPath = wsPath.startsWith('/') ? wsPath : `/${wsPath}`
-    const protocol = window.location.protocol === 'https:' ? 'wss:' : 'ws:'
-    return `${protocol}//${window.location.host}${normalizedPath}`
-  }
-
-  const websocketUrl = computed<string>(() => resolveWebSocketUrl(configuredBackendUrl.value))
-
   const logDebug = (...args: unknown[]): void => {
     if (import.meta.env.DEV) {
       console.info('[GenerationStudio]', ...args)
     }
   }
 
-  const parseTimestamp = (value?: string): number => {
-    if (!value) {
-      return 0
-    }
-    const timestamp = Date.parse(value)
-    return Number.isNaN(timestamp) ? 0 : timestamp
+  const showToast = (message: string, type: NotificationType = 'success'): void => {
+    logDebug(`[${type.toUpperCase()}] ${message}`)
+    appStore.addNotification(message, type)
   }
 
-  const normalizeProgress = (value?: number | null): number => {
-    if (typeof value !== 'number' || Number.isNaN(value)) {
-      return 0
-    }
-    return value <= 1 ? Math.round(value * 100) : Math.round(value)
-  }
-
-  const isGenerationProgressMessage = (
-    message: WebSocketMessage,
-  ): message is GenerationProgressMessage => message.type === 'generation_progress'
-
-  const isGenerationCompleteMessage = (
-    message: WebSocketMessage,
-  ): message is GenerationCompleteMessage => message.type === 'generation_complete'
-
-  const isGenerationErrorMessage = (
-    message: WebSocketMessage,
-  ): message is GenerationErrorMessage => message.type === 'generation_error'
-
-  const { fetchData: loadSystemStatus } = useSystemStatusApi()
-  const { fetchData: loadActiveJobsData } = useActiveJobsApi()
-  const { fetchData: loadRecentResultsData } = useRecentResultsApi(() => {
-    const limit = showHistory.value ? 50 : 10
-    return resolveBackendUrl(`/generation/results?limit=${limit}`, configuredBackendUrl.value)
+  const { loadSavedParams, saveParams, savePreset, loadFromComposer, useRandomPrompt } = useGenerationPersistence({
+    params,
+    showToast,
   })
 
-  const sortedActiveJobs: ComputedRef<GenerationJob[]> = computed(() => {
-    const statusPriority: Record<string, number> = {
-      processing: 0,
-      queued: 1,
-      completed: 2,
-      failed: 3,
-    }
-
-    return [...activeJobs.value].sort((a, b) => {
-      const aPriority = statusPriority[a.status] ?? 4
-      const bPriority = statusPriority[b.status] ?? 4
-
-      if (aPriority !== bPriority) {
-        return aPriority - bPriority
-      }
-
-      const aCreated = parseTimestamp(a.created_at ?? a.startTime)
-      const bCreated = parseTimestamp(b.created_at ?? b.startTime)
-
-      return bCreated - aCreated
-    })
+  const {
+    activeJobs,
+    recentResults,
+    sortedActiveJobs,
+    isConnected,
+    initialize: initializeUpdates,
+    loadRecentResultsData,
+  } = useGenerationUpdates({
+    appStore,
+    systemStatus,
+    showHistory,
+    configuredBackendUrl,
+    logDebug,
+    showToast,
   })
-
-  const loadSystemStatusData = async (): Promise<void> => {
-    try {
-      const result = await loadSystemStatus()
-      if (result) {
-        const payload = result as SystemStatusPayload
-        const { metrics: _metrics, message: _message, updated_at: _updatedAt, ...status } = payload
-        systemStatus.value = {
-          ...systemStatus.value,
-          ...(status as Partial<SystemStatusState>),
-        }
-        appStore.updateSystemStatus(status as Partial<SystemStatusState>)
-      }
-    } catch (error) {
-      console.error('Failed to load system status:', error)
-    }
-  }
-
-  const loadActiveJobsDataFn = async (): Promise<void> => {
-    try {
-      const result = await loadActiveJobsData()
-      if (Array.isArray(result)) {
-        appStore.setActiveJobs(result)
-      }
-    } catch (error) {
-      console.error('Failed to load active jobs:', error)
-    }
-  }
-
-  const loadRecentResultsDataFn = async (): Promise<void> => {
-    try {
-      const result = await loadRecentResultsData()
-      if (Array.isArray(result)) {
-        appStore.setRecentResults(result)
-      }
-    } catch (error) {
-      console.error('Failed to load recent results:', error)
-    }
-  }
-
-  const handleWebSocketMessage = (event: MessageEvent): void => {
-    try {
-      const data = JSON.parse(event.data as string) as WebSocketMessage
-      if (!data || typeof data !== 'object') {
-        logDebug('Received invalid WebSocket message:', data)
-        return
-      }
-
-      switch (data.type) {
-        case 'generation_progress':
-          if (isGenerationProgressMessage(data)) {
-            updateJobProgress(data)
-          }
-          break
-        case 'generation_complete':
-          if (isGenerationCompleteMessage(data)) {
-            handleGenerationComplete(data)
-          }
-          break
-        case 'generation_error':
-          if (isGenerationErrorMessage(data)) {
-            handleGenerationError(data)
-          }
-          break
-        case 'queue_update':
-          if (Array.isArray(data.jobs)) {
-            appStore.setActiveJobs(data.jobs)
-          }
-          break
-        case 'system_status': {
-          const { metrics: _metrics, message: _message, updated_at: _updatedAt, type: _type, ...status } = data
-          systemStatus.value = {
-            ...systemStatus.value,
-            ...(status as Partial<SystemStatusState>),
-          }
-          appStore.updateSystemStatus(status as Partial<SystemStatusState>)
-          break
-        }
-        case 'generation_started':
-          logDebug('Generation job started', data.job_id)
-          break
-        default:
-          logDebug('Unknown WebSocket message type:', (data as { type?: string }).type)
-      }
-    } catch (error) {
-      console.error('Failed to parse WebSocket message:', error)
-    }
-  }
-
-  const initWebSocket = (): void => {
-    try {
-      const wsUrl = websocketUrl.value
-
-      if (websocket.value) {
-        websocket.value.onclose = null
-        websocket.value.close()
-      }
-
-      const connection = new WebSocket(wsUrl)
-      websocket.value = connection
-
-      connection.onopen = () => {
-        logDebug('WebSocket connected for generation updates')
-      }
-
-      connection.onmessage = handleWebSocketMessage
-
-      connection.onerror = (event) => {
-        console.error('WebSocket error:', event)
-      }
-
-      connection.onclose = () => {
-        logDebug('WebSocket connection closed')
-        if (websocket.value === connection) {
-          websocket.value = null
-        }
-        if (typeof window !== 'undefined') {
-          window.setTimeout(() => {
-            if (!websocket.value) {
-              initWebSocket()
-            }
-          }, 3000)
-        }
-      }
-    } catch (error) {
-      console.error('Failed to initialize WebSocket:', error)
-    }
-  }
-
-  const startPolling = (): void => {
-    if (pollInterval.value) {
-      return
-    }
-    pollInterval.value = window.setInterval(async () => {
-      if (activeJobs.value.length > 0) {
-        await loadActiveJobsDataFn()
-      }
-      await loadSystemStatusData()
-    }, 2000)
-  }
-
-  const stopPolling = (): void => {
-    if (pollInterval.value != null) {
-      window.clearInterval(pollInterval.value)
-      pollInterval.value = null
-    }
-  }
 
   const startGeneration = async (): Promise<void> => {
     const trimmedPrompt = params.value.prompt.trim()
@@ -367,54 +150,6 @@ export const useGenerationStudio = () => {
     await Promise.allSettled(cancellableJobs.map((job) => cancelJob(job.id)))
   }
 
-  const updateJobProgress = (update: ProgressUpdate): void => {
-    const job = activeJobs.value.find((item) => item.id === update.job_id)
-    if (!job) {
-      return
-    }
-
-    job.progress = normalizeProgress(update.progress)
-    job.status = update.status as GenerationJob['status']
-
-    if (typeof update.current_step === 'number') {
-      job.current_step = update.current_step
-    }
-
-    if (typeof update.total_steps === 'number') {
-      job.total_steps = update.total_steps
-    }
-  }
-
-  const handleGenerationComplete = (data: Extract<WebSocketMessage, { type: 'generation_complete' }>): void => {
-    appStore.removeJob(data.job_id)
-
-    const createdAt = data.created_at ?? new Date().toISOString()
-    const imageUrl = data.image_url ?? (Array.isArray(data.images) ? data.images[0] ?? null : null)
-
-    const result: GenerationResult = {
-      id: data.result_id ?? data.job_id,
-      job_id: data.job_id,
-      result_id: data.result_id,
-      prompt: data.prompt,
-      negative_prompt: data.negative_prompt,
-      image_url: imageUrl,
-      width: data.width,
-      height: data.height,
-      steps: data.steps,
-      cfg_scale: data.cfg_scale,
-      seed: data.seed ?? null,
-      created_at: createdAt,
-    }
-
-    appStore.addResult(result)
-    showToast('Generation completed successfully', 'success')
-  }
-
-  const handleGenerationError = (data: Extract<WebSocketMessage, { type: 'generation_error' }>): void => {
-    appStore.removeJob(data.job_id)
-    showToast(`Generation failed: ${data.error}`, 'error')
-  }
-
   const showImageModal = (result: GenerationResult | null): void => {
     if (!result) {
       return
@@ -469,97 +204,8 @@ export const useGenerationStudio = () => {
   }
 
   const refreshResults = async (): Promise<void> => {
-    await loadRecentResultsDataFn()
+    await loadRecentResultsData()
     showToast('Results refreshed', 'success')
-  }
-
-  const loadFromComposer = (): void => {
-    try {
-      const composerData = localStorage.getItem('composerPrompt')
-      if (composerData) {
-        params.value.prompt = composerData
-        showToast('Loaded prompt from composer', 'success')
-      } else {
-        showToast('No composer data found', 'warning')
-      }
-    } catch (error) {
-      console.error('Error loading composer data:', error)
-    }
-  }
-
-  const useRandomPrompt = (): void => {
-    const randomPrompts: readonly string[] = [
-      'a beautiful anime girl with long flowing hair',
-      'a majestic dragon soaring through cloudy skies',
-      'a cyberpunk cityscape with neon lights',
-      'a serene landscape with mountains and a lake',
-      'a cute robot in a futuristic laboratory',
-      'a magical forest with glowing mushrooms',
-      'a space station orbiting a distant planet',
-      'a steampunk airship flying over Victorian city',
-    ]
-
-    const index = Math.floor(Math.random() * randomPrompts.length)
-    params.value.prompt = randomPrompts[index]
-    showToast('Random prompt generated', 'success')
-  }
-
-  const savePreset = (): void => {
-    const presetName = window.prompt('Enter a name for this preset:')
-    if (!presetName) {
-      return
-    }
-
-    const preset = {
-      name: presetName,
-      params: { ...params.value },
-      created_at: new Date().toISOString(),
-    }
-
-    try {
-      const savedPresets = JSON.parse(localStorage.getItem('generationPresets') ?? '[]') as unknown[]
-      savedPresets.push(preset)
-      localStorage.setItem('generationPresets', JSON.stringify(savedPresets))
-      showToast(`Preset "${presetName}" saved`, 'success')
-    } catch (error) {
-      console.error('Failed to save preset:', error)
-      showToast('Failed to save preset', 'error')
-    }
-  }
-
-  const loadSavedParams = (): void => {
-    try {
-      const urlParams = new URLSearchParams(window.location.search)
-      const prompt = urlParams.get('prompt')
-      if (typeof prompt === 'string') {
-        params.value.prompt = prompt
-      }
-
-      const saved = localStorage.getItem('generation_params')
-      if (saved) {
-        const parsed = JSON.parse(saved) as Partial<GenerationFormState>
-        Object.assign(params.value, parsed)
-      }
-    } catch (error) {
-      console.error('Error loading saved parameters:', error)
-    }
-  }
-
-  const saveParams = (value: GenerationFormState = params.value): void => {
-    try {
-      localStorage.setItem('generation_params', JSON.stringify(value))
-    } catch (error) {
-      console.error('Error saving parameters:', error)
-    }
-  }
-
-  const cleanup = (): void => {
-    if (websocket.value) {
-      websocket.value.onclose = null
-      websocket.value.close()
-      websocket.value = null
-    }
-    stopPolling()
   }
 
   const formatTime = (dateString?: string): string => {
@@ -628,59 +274,10 @@ export const useGenerationStudio = () => {
     }
   }
 
-  const showToast = (message: string, type: NotificationType = 'success'): void => {
-    logDebug(`[${type.toUpperCase()}] ${message}`)
-    appStore.addNotification(message, type)
-  }
-
-  watch(showHistory, () => {
-    void loadRecentResultsDataFn()
-  })
-
-  watch(params, (newParams) => {
-    saveParams(newParams)
-  }, { deep: true })
-
-  watch(isConnected, (connected) => {
-    logDebug('WebSocket connection state changed:', connected)
-  })
-
-  watch(websocketUrl, (newUrl, oldUrl) => {
-    if (!newUrl || newUrl === oldUrl) {
-      return
-    }
-
-    if (websocket.value) {
-      const currentConnection = websocket.value
-      currentConnection.onclose = null
-      try {
-        currentConnection.close()
-      } catch (error) {
-        console.error('Failed to close existing WebSocket connection:', error)
-      } finally {
-        websocket.value = null
-      }
-    }
-
-    initWebSocket()
-  })
-
   onMounted(async () => {
     logDebug('Initializing Generation Studio composable...')
-
-    await Promise.all([
-      loadSystemStatusData(),
-      loadActiveJobsDataFn(),
-      loadRecentResultsDataFn(),
-    ])
-
-    initWebSocket()
-    startPolling()
+    await initializeUpdates()
     loadSavedParams()
-  })
-
-  onUnmounted(() => {
-    cleanup()
   })
 
   return {
@@ -693,6 +290,7 @@ export const useGenerationStudio = () => {
     activeJobs,
     recentResults,
     sortedActiveJobs,
+    isConnected,
     startGeneration,
     cancelJob,
     clearQueue,
